# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin rbac manage API",
        "title": "gin rbac",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "API_HOST:49000",
    "basePath": "/",
    "paths": {
...
        "/rbac/v1/action/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 功能配置"
                ],
                "summary": "[功能配置删除]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DelActionConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/ginutil.BaseResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/action/find": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据功能ID，查询基础信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 功能配置"
                ],
                "summary": "[功能配置查询]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FindActionConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/model.FindActionConfigResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/action/import": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "导入Swagger JSON 文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 功能配置"
                ],
                "summary": "[功能配置导入Swagger]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SwaggerJSONToActionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/ginutil.BaseResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/action/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 功能配置"
                ],
                "summary": "[功能配置列表]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ListActionConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/model.ListActionConfig"
                        }
                    }
                }
            }
        },
        "/rbac/v1/action/upsert": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 功能配置"
                ],
                "summary": "[功能配置创建/更新]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpsertActionConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/ginutil.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/rbac/v1/menu/action/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 菜单配置"
                ],
                "summary": "[菜单配置功能更新]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateMenuConfigActionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/ginutil.BaseResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/menu/actions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 菜单配置"
                ],
                "summary": "[菜单配置功能]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetMenuActionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/model.GetMenuActionsResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/menu/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 菜单配置"
                ],
                "summary": "[菜单配置创建]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateMenuConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/ginutil.BaseResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/menu/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 菜单配置"
                ],
                "summary": "[菜单配置删除]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DelMenuConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/ginutil.BaseResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/menu/tree": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 菜单配置"
                ],
                "summary": "[菜单配置树]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetMenuTreeDirsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/model.GetMenuTreeDirsResp"
                        }
                    }
                }
            }
        },
        "/rbac/v1/menu/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC 菜单配置"
                ],
                "summary": "[菜单配置更新]",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateMenuConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/ginutil.BaseResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ginutil.BaseResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "model.Action": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.ActionBase": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.CreateMenuConfigReq": {
            "type": "object",
            "required": [
                "appId",
                "name",
                "path",
                "typ"
            ],
            "properties": {
                "appId": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                },
                "typ": {
                    "type": "integer"
                }
            }
        },
...
        "model.DelActionConfigReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "model.DelMenuConfigReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.FindActionConfigReq": {
            "type": "object",
            "required": [
                "appId"
            ],
            "properties": {
                "actionId": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "appId": {
                    "type": "string"
                }
            }
        },
        "model.FindActionConfigResp": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ActionBase"
                    }
                }
            }
        },
        "model.GetMenuActionsReq": {
            "type": "object",
            "required": [
                "menuId"
            ],
            "properties": {
                "menuId": {
                    "type": "integer"
                }
            }
        },
        "model.GetMenuActionsResp": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Action"
                    }
                }
            }
        },
        "model.GetMenuTreeDirsReq": {
            "type": "object",
            "required": [
                "appId"
            ],
            "properties": {
                "appId": {
                    "type": "string"
                }
            }
        },
        "model.GetMenuTreeDirsResp": {
            "type": "object",
            "properties": {
                "dirs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MenuTreeDir"
                    }
                }
            }
        },
...
        "model.ListActionConfig": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "integer"
                }
            }
        },
        "model.ListActionConfigReq": {
            "type": "object",
            "required": [
                "page",
                "size"
            ],
            "properties": {
                "appId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "model.MenuTreeDir": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "appId": {
                    "type": "string"
                },
                "children": {
                    "$ref": "#/definitions/model.MenuTreeDirs"
                },
                "id": {
                    "type": "integer"
                },
                "memo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "typ": {
                    "type": "integer"
                }
            }
        },
        "model.MenuTreeDirs": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/model.MenuTreeDir"
            }
        },
...
        "model.SwaggerJSONToActionsReq": {
            "type": "object",
            "required": [
                "appId"
            ],
            "properties": {
                "appId": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                },
                "swaggerTxt": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateMenuConfigActionReq": {
            "type": "object",
            "required": [
                "menuId"
            ],
            "properties": {
                "actionId": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "menuId": {
                    "type": "integer"
                }
            }
        },
        "model.UpdateMenuConfigReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "memo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                },
                "typ": {
                    "type": "integer"
                }
            }
        },
...
        "model.UpsertActionConfigReq": {
            "type": "object",
            "required": [
                "appId",
                "method",
                "name",
                "path",
                "status"
            ],
            "properties": {
                "appId": {
                    "type": "string"
                },
                "method": {
                    "description": "GET POST PUT DELETE",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "description": "1-正常 2-锁定",
                    "type": "integer"
                }
            }
        },
...
    }
}